+++ src/Language/Brainfuck/Interpreter.hs
@@ -6,15 +6,20 @@ module Language.Brainfuck.Interpreter
 where
 
 import qualified Data.IntMap.Strict            as M
+import           Data.Maybe
+import           Control.Exception
+import           Control.Monad
+import           Control.Monad.IO.Class
 
-import           Language.Brainfuck.Parser
 import           Language.Brainfuck.Machine
 import           Language.Brainfuck.Optimizer
-import           Control.Monad.IO.Class
-import           Control.Monad
+import           Language.Brainfuck.Parser
 
-convertEnum :: (Enum a, Enum b) => a -> b
-convertEnum = toEnum . fromEnum
+toCharThrow :: BFCell t => t -> Char
+toCharThrow n = fromMaybe (throw BadToCharConv) (toChar n)
+
+fromCharThrow :: BFCell t => Char -> t
+fromCharThrow c = fromMaybe (throw (BadFromCharConv c)) (fromChar c)
 
 interpretBasic :: BFCell t => [BF] -> BrainfuckM t ()
 interpretBasic = mapM_ interpInstr  where
@@ -22,8 +27,8 @@ interpretBasic = mapM_ interpInstr  where
     interpInstr Decrement   = modifyCell (\n -> n - 1)
     interpInstr MoveLeft    = moveHead (-1)
     interpInstr MoveRight   = moveHead 1
-    interpInstr Input       = liftIO getChar >>= setCell . convertEnum
-    interpInstr Output      = readCell >>= liftIO . putChar . convertEnum
+    interpInstr Input       = liftIO getChar >>= setCell . fromCharThrow
+    interpInstr Output      = readCell >>= liftIO . putChar . toCharThrow
     interpInstr (Loop code) = do
         cell <- readCell
         unless (cell == 0) (interpretBasic code >> interpInstr (Loop code))
@@ -33,8 +38,8 @@ interpretCollapsed :: BFCell t => [CollapsedBF t] -> BrainfuckM t ()
 interpretCollapsed = mapM_ interpInstr  where
     interpInstr (Incr n)          = modifyCell (+ n)
     interpInstr (Move n)          = moveHead n
-    interpInstr CInput            = liftIO getChar >>= setCell . convertEnum
-    interpInstr COutput           = readCell >>= liftIO . putChar . convertEnum
+    interpInstr CInput            = liftIO getChar >>= setCell . fromCharThrow
+    interpInstr COutput           = readCell >>= liftIO . putChar . toCharThrow
     interpInstr loop@(CLoop code) = do
         cell <- readCell
         unless (cell == 0) (interpretCollapsed code >> interpInstr loop)
@@ -44,8 +49,8 @@ interpretOpt :: BFCell t => [OptBF t] -> BrainfuckM t ()
 interpretOpt = mapM_ interpInstr  where
     interpInstr (OIncr n)         = modifyCell (+ n)
     interpInstr (OMove n)         = moveHead n
-    interpInstr OInput            = liftIO getChar >>= setCell . convertEnum
-    interpInstr OOutput           = readCell >>= liftIO . putChar . convertEnum
+    interpInstr OInput            = liftIO getChar >>= setCell . fromCharThrow
+    interpInstr OOutput           = readCell >>= liftIO . putChar . toCharThrow
     interpInstr loop@(OLoop code) = do
         cell <- readCell
         unless (cell == 0) (interpretOpt code >> interpInstr loop)